Add 'is_admin' column to User table for CRUD on Categories/clues
Add 'is_admin' column which is set to 'false' by default. If a user
signs up with 'admin@admin,' then it is set to true for that user.

Refactor Category and Clue Update/Delete actions to limit them to
only admins, so only the admin can delete or update a category or clue
through a CURL request.