Hopefully fix category population bug
Refactor seed algorithm to seed each category with only
five clues, and change the 'create_game' function to make
a game_clue for each one of those categories.

This is scaled down and removes some of the 'randomness' functionality
I had included previously, but on the brightside, it works now.